refactor: rework examples and move them to workspace examples
