+++ crates/core/Cargo.toml
@@ -19,7 +19,7 @@ publish = true
 readme = "README.md"
 repository = "https://github.com/rustic-rs/rustic_core"
 resolver = "2"
-rust-version = "1.71.1"
+rust-version = { workspace = true }
 description = """
 rustic_core - library for fast, encrypted, deduplicated backups that powers rustic-rs
 """
@@ -81,22 +81,35 @@ serde_with = { version = "3.4.0", features = ["base64"] }
 
 # other dependencies
 bytes = "1.5.0"
-chrono = { version = "0.4.31", default-features = false, features = ["clock", "serde"] }
+chrono = { version = "0.4.31", default-features = false, features = [
+  "clock",
+  "serde",
+] }
 enum-map = "2.7.3"
 enum-map-derive = "0.17.0"
 zstd = "0.13.0"
 
 # local backend
 aho-corasick = "1.1.2"
-cached = { version = "0.46.1", default-features = false, features = ["proc_macro"] }
+cached = { version = "0.46.1", default-features = false, features = [
+  "proc_macro",
+] }
 filetime = "0.2.23"
 ignore = "0.4.21"
-nix = { version = "0.27.1", default-features = false, features = ["user", "fs"] }
+nix = { version = "0.27.1", default-features = false, features = [
+  "user",
+  "fs",
+] }
 walkdir = "2.4.0"
 
 # rest backend
 backoff = "0.4.0"
-reqwest = { version = "0.11.23", default-features = false, features = ["json", "rustls-tls-native-roots", "stream", "blocking"] }
+reqwest = { version = "0.11.23", default-features = false, features = [
+  "json",
+  "rustls-tls-native-roots",
+  "stream",
+  "blocking",
+] }
 url = "2.5.0"
 
 # cache
@@ -104,7 +117,11 @@ cachedir = "0.3.1"
 dirs = "5.0.1"
 
 # cli
-clap = { version = "4.4.11", optional = true, features = ["derive", "env", "wrap_help"] }
+clap = { version = "4.4.11", optional = true, features = [
+  "derive",
+  "env",
+  "wrap_help",
+] }
 
 anyhow = "1.0.75"
 bytesize = "1.3.0"
@@ -119,7 +136,11 @@ shell-words = "1.1.0"
 
 [target.'cfg(not(windows))'.dependencies]
 # opendal backend - sftp is not supported on windows, see https://github.com/apache/incubator-opendal/issues/2963
-opendal = { version = "0.43.0", features = ["services-b2", "services-sftp", "services-swift"] }
+opendal = { version = "0.43.0", features = [
+  "services-b2",
+  "services-sftp",
+  "services-swift",
+] }
 sha2 = { version = "0.10", features = ["asm"] }
 
 [target.'cfg(windows)'.dependencies]
