--- examples/backup_example/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "backup_example"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rustic_backend = { path = "../../crates/backends" }
-rustic_core = { path = "../../crates/core" }
-simplelog = "0.12.1"
