+++ examples/check/examples/check.rs
@@ -1,0 +1,24 @@
+//! `check` example
+use rustic_backend::BackendOptions;
+use rustic_core::{CheckOptions, Repository, RepositoryOptions};
+use simplelog::{Config, LevelFilter, SimpleLogger};
+use std::error::Error;
+
+fn main() -> Result<(), Box<dyn Error>> {
+    // Display info logs
+    let _ = SimpleLogger::init(LevelFilter::Info, Config::default());
+
+    // Initialize Backends
+    let backends = BackendOptions::default()
+        .repository("/tmp/repo")
+        .to_backends()?;
+
+    // Open repository
+    let repo_opts = RepositoryOptions::default().password("test");
+    let repo = Repository::new(&repo_opts, backends)?.open()?;
+
+    // Check repository with standard options but omitting cache checks
+    let opts = CheckOptions::default().trust_cache(true);
+    repo.check(opts)?;
+    Ok(())
+}
