+++ examples/check/src/lib.rs
