+++ examples/prune/src/lib.rs
@@ -1,0 +1,1 @@
+
