+++ examples/tag/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tag"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rustic_backend = { workspace = true }
+rustic_core = { workspace = true }
+simplelog = { workspace = true }
